[DEV][kconfig] Fix warnings.

+++ kconfig/lxdialog/textbox.c
@@ -343,10 +343,10 @@ static void print_line(WINDOW * win, int row, int width)
 	char *line;
 
 	line = get_line();
-	line += MIN(strlen(line), hscroll);	/* Scroll horizontally */
+	line += MIN((int)strlen(line), hscroll);	/* Scroll horizontally */
 	wmove(win, row, 0);	/* move cursor to correct line */
 	waddch(win, ' ');
-	waddnstr(win, line, MIN(strlen(line), width - 2));
+	waddnstr(win, line, MIN((int)strlen(line), width - 2));
 
 	/* Clear 'residue' of previous line */
 #if OLD_NCURSES
