+++ kconfig/mconf.c
@@ -301,12 +301,12 @@ static void set_config_filename(const char *config_filename)
 
 	size = snprintf(menu_backtitle, sizeof(menu_backtitle),
 			"%s - %s", config_filename, rootmenu.prompt->text);
-	if (size >= sizeof(menu_backtitle))
+	if (size >= (int)sizeof(menu_backtitle))
 		menu_backtitle[sizeof(menu_backtitle)-1] = '\0';
 	set_dialog_backtitle(menu_backtitle);
 
 	size = snprintf(filename, sizeof(filename), "%s", config_filename);
-	if (size >= sizeof(filename))
+	if (size >= (int)sizeof(filename))
 		filename[sizeof(filename)-1] = '\0';
 }
